requirements.txt
awscli
aws-mfa
pytest
click
pyhcl
pytz
PyYAML
tzlocal
ansible
terraform_external_data
python-jenkins
jenkinsapi
